refactor(removeSeparator): split str by separators and return sliced str list
feature #2.4.
- 남은 문자열을 커스텀 구분자를 기준으로 자르고, 리스트로 반환하는 메서드로 수정했습니다

Closes "issue: 남은 문자열을 커스텀 구분자를 기준으로 자르고, 리스트로 반환한다"
